requirements.txt
boto3==1.39.8
fastapi==0.115.12
httpx==0.27.2
opensearch-py==2.7.1
pydantic==2.11.3
python-dotenv==1.0.1
pytest==8.3.5
requests-aws4auth==1.3.1
streamlit==1.44.1
uvicorn==0.34.0
